Refactor to Javascript
