+++ .eslintrc.js
@@ -1,19 +1,17 @@
 module.exports = {
-  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": 2018,
     "sourceType": "module",
-    "project": "./tsconfig.json",
   },
   "extends": [
-    "plugin:@typescript-eslint/recommended",
-    "prettier",
-    "prettier/@typescript-eslint"
+    "airbnb",
   ],
   "env": {
       "es6": true,
       "node": true,
       "jest": true,
   },
-  "plugins": ["@typescript-eslint"],
+  "rules": {
+    "max-len": false,
+  }
 };
