+++ index.js
@@ -1,6 +1,12 @@
 
-import fs from 'fs';
-import { promisify } from 'util';
+const fs = require('fs');
+const { promisify } = require('util');
+const cloneDeep = require('lodash/cloneDeep');
+const isPlainObject = require('lodash/isPlainObject');
+const uuidv4 = require('uuid/v4');
+const ms = require('ms');
+const tinydate = require('tinydate');
+const MessagePack = require('what-the-pack');
 
 const writeFile = promisify(fs.writeFile);
 const open = promisify(fs.open);
@@ -12,45 +18,16 @@ const fstat = promisify(fs.fstat);
 const exists = promisify(fs.exists);
 const mkdir = promisify(fs.mkdir);
 
-import cloneDeep from 'lodash/cloneDeep';
-import isPlainObject from 'lodash/isPlainObject';
-import uuidv4 from 'uuid/v4';
-import ms from 'ms';
-
-import tinydate from 'tinydate';
-import MessagePack from 'what-the-pack';
-
-export type AcceptedValues = string|number|boolean|null|undefined;
-
-export interface Item {
-  readonly id ?: string;
-  [key:string]: AcceptedValues|(AcceptedValues)[];
-}
+// Database Version
+const rurudbVersion = 9;
 
 // Helper Functions
 const dateToString = tinydate('{DD}-{MM}-{YY}-{HH}-{mm}-{ss}');
-const compareString = (
-  a: string,
-  b: string,
-  descend: boolean
-) : number => (descend ? b.localeCompare(a) : a.localeCompare(b));
-const compareNumber = (
-  a: number,
-  b: number,
-  descend: boolean
-) : number => (descend ? b - a : a - b);
+const compareString = (a, b, descend) => (descend ? b.localeCompare(a) : a.localeCompare(b));
+const compareNumber = (a, b, descend) => (descend ? b - a : a - b);
 
-export class Query <ExtendedItem extends Item> {
-  private resultIds: string[];
-  private resultItems: Item[];
-  private slicedItems: Item[];
-  private queryOffset: number;
-  private queryLimit: number;
-  private sorts: ([string, boolean]|[Function])[];
-  private selectedFields: string[];
-  private hiddenFields: string[];
-  private finalized: boolean;
-  public constructor (table: Table <ExtendedItem> ) {
+class Query {
+  constructor(table) {
     if (typeof table !== 'object') {
       throw Error('@constructor : Expecting "table" to be typeof "object"');
     }
@@ -64,135 +41,154 @@ export class Query <ExtendedItem extends Item> {
     this.hiddenFields = [];
     this.finalized = false;
   }
-  public offset (value: number) : Query <ExtendedItem> {
+
+  offset(value) {
     if (this.finalized) throw Error('@offset : Query must not be finalized yet');
     this.queryOffset = value;
     return this;
   }
-  public limit (value: number) : Query <ExtendedItem> {
+
+  limit(value) {
     if (this.finalized) throw Error('@limit : Query must not be finalized yet');
     this.queryLimit = value;
     return this;
   }
-  public ascend (field: string) : Query <ExtendedItem> {
+
+  ascend(field) {
     if (this.finalized) throw Error('@ascend : Query must not be finalized yet');
     this.sorts.push([field, false]);
     return this;
   }
-  public descend (field: string) : Query <ExtendedItem> {
+
+  descend(field) {
     if (this.finalized) throw Error('@descend : Query must not be finalized yet');
     this.sorts.push([field, true]);
     return this;
   }
-  public sortBy (sortFn: (a: Item, b: Item) => number) : Query <ExtendedItem> {
+
+  sortBy(sortFn) {
     if (this.finalized) throw Error('@sortBy : Query must not be finalized yet');
     this.sorts.push([sortFn]);
     return this;
   }
-  public gt (field: string, value: number) : Query <ExtendedItem> {
+
+  gt(field, value) {
     if (this.finalized) throw Error('@gt : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field] as number) && item[field] as number > value);
+    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field]) && item[field] > value);
     return this;
   }
-  public gte (field: string, value: number) : Query <ExtendedItem> {
+
+  gte(field, value) {
     if (this.finalized) throw Error('@gte : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field] as number) && item[field] as number >= value);
+    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field]) && item[field] >= value);
     return this;
   }
-  public lt (field: string, value: number) : Query <ExtendedItem> {
-    if (this.finalized) throw Error('@lt : Query must not be finalized yet');  
-    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field] as number) && item[field] as number < value);
+
+  lt(field, value) {
+    if (this.finalized) throw Error('@lt : Query must not be finalized yet');
+    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field]) && item[field] < value);
     return this;
   }
-  public lte (field: string, value: number) : Query <ExtendedItem> {
+
+  lte(field, value) {
     if (this.finalized) throw Error('@lte : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field] as number) && item[field] as number <= value);
+    this.slicedItems = this.slicedItems.filter(item => Number.isFinite(item[field]) && item[field] <= value);
     return this;
   }
-  public eq (field: string, value: number) : Query <ExtendedItem> {
+
+  eq(field, value) {
     if (this.finalized) throw Error('@eq : Query must not be finalized yet');
     this.slicedItems = this.slicedItems.filter(item => item[field] === value);
     return this;
   }
-  public neq (field: string, value: number) : Query <ExtendedItem> {
+
+  neq(field, value) {
     if (this.finalized) throw Error('@neq : Query must not be finalized yet');
     this.slicedItems = this.slicedItems.filter(item => item[field] !== value);
     return this;
   }
-  public has (field: string, value: (AcceptedValues)) : Query <ExtendedItem> {
+
+  has(field, value) {
     if (this.finalized) throw Error('@has : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && (item[field] as (AcceptedValues)[]).includes(value));
+    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && (item[field]).includes(value));
     return this;
   }
-  public hasAnyOf (field: string, values: (AcceptedValues)[]) : Query <ExtendedItem> {
+
+  hasAnyOf(field, values) {
     if (this.finalized) throw Error('@hasAnyOf : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.some(value => (item[field] as (AcceptedValues)[]).includes(value)));
+    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.some(value => (item[field]).includes(value)));
     return this;
   }
-  public hasAllOf (field: string, values: (AcceptedValues)[]) : Query <ExtendedItem> {
+
+  hasAllOf(field, values) {
     if (this.finalized) throw Error('@hasAllOf : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.every(value => (item[field] as (AcceptedValues)[]).includes(value)));
+    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.every(value => (item[field]).includes(value)));
     return this;
   }
-  public hasNoneOfAny (field: string, values: (AcceptedValues)[]) : Query <ExtendedItem> {
-    if (this.finalized) throw Error('@hasNoneOfAny : Query must not be finalized yet');  
-    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.some(value => (item[field] as (AcceptedValues)[]).includes(value) === false));
+
+  hasNoneOfAny(field, values) {
+    if (this.finalized) throw Error('@hasNoneOfAny : Query must not be finalized yet');
+    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.some(value => (item[field]).includes(value) === false));
     return this;
   }
-  public hasNoneOfAll (field: string, values: (AcceptedValues)[]) : Query <ExtendedItem> {
+
+  hasNoneOfAll(field, values) {
     if (this.finalized) throw Error('@hasNoneOfAll : Query must not be finalized yet');
-    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.every(value => (item[field] as (AcceptedValues)[]).includes(value) === false));
+    this.slicedItems = this.slicedItems.filter(item => Array.isArray(item[field]) && values.every(value => (item[field]).includes(value) === false));
     return this;
   }
-  public select (...fields: string[]) : Query <ExtendedItem> {
+
+  select(...fields) {
     if (this.finalized) throw Error('@select : Query must not be finalized yet');
     this.selectedFields = fields.slice();
     return this;
   }
-  public hide (...fields: string[]) : Query <ExtendedItem> {
+
+  hide(...fields) {
     if (this.finalized) throw Error('@hide : Query must not be finalized yet');
     this.hiddenFields = fields.slice();
     return this;
   }
-  public filterBy (filterFn: (item: Item) => boolean) : Query <ExtendedItem> {
+
+  filterBy(filterFn) {
     if (this.finalized) throw Error('@filterBy : Query must not be finalized yet');
     this.slicedItems = this.slicedItems.filter(item => filterFn(item));
     return this;
   }
-  private finalize () : void {
+
+  finalize() {
     if (this.sorts.length > 0) {
       this.slicedItems.sort((a, b) => {
         for (let i = 0, l = this.sorts.length; i < l; i += 1) {
           const sort = this.sorts[i];
           if (typeof sort[0] === 'function') {
-            const [sortFn] = sort as [Function];            
+            const [sortFn] = sort;
             return sortFn(a, b);
-          } else {
-            const [field, fieldDescend] = sort as [string, boolean];
-            if (typeof a[field] !== typeof b[field]) { // If field of both slicedItems don't match: EXIT LOOP
-              break;
-            } else if (typeof a[field] !== 'string' && typeof a[field] !== 'number') { // If both item fields are't "string" or "number": EXIT LOOP
-              break;
-            } else if (a[field] === b[field]) { // If value of both slicedItems are equal: SKIP SORT
-              continue;
-            } else if (typeof a[field] === 'string') {
-              return compareString(a[field] as string, b[field] as string, fieldDescend as boolean);            
-            } else if (typeof a[field] === 'number') {
-              return compareNumber(a[field] as number, b[field] as number, fieldDescend as boolean);
-            }
+          }
+          const [field, fieldDescend] = sort;
+          if (typeof a[field] !== typeof b[field]) { // If field of both slicedItems don't match: EXIT LOOP
+            break;
+          } else if (typeof a[field] !== 'string' && typeof a[field] !== 'number') { // If both item fields are't "string" or "number": EXIT LOOP
+            break;
+          } else if (a[field] === b[field]) { // If value of both slicedItems are equal: SKIP SORT
+            continue;
+          } else if (typeof a[field] === 'string') {
+            return compareString(a[field], b[field], fieldDescend);
+          } else if (typeof a[field] === 'number') {
+            return compareNumber(a[field], b[field], fieldDescend);
           }
         }
         return 0;
       });
     }
     this.slicedItems = this.slicedItems.slice(this.queryOffset, this.queryOffset + this.queryLimit);
-    const resultIds: string[] = new Array(this.slicedItems.length);
-    const resultItems: Item[] = new Array(this.slicedItems.length);
+    const resultIds = new Array(this.slicedItems.length);
+    const resultItems = new Array(this.slicedItems.length);
     for (let i = 0, l = this.slicedItems.length; i < l; i += 1) {
-      const item = cloneDeep (this.slicedItems[i]);
-      resultIds[i] = item.id as string;     
+      const item = cloneDeep(this.slicedItems[i]);
+      resultIds[i] = item.id;
       if (this.selectedFields.length > 0) {
-        const itemFields = Object.keys(item); 
+        const itemFields = Object.keys(item);
         for (let a = 0, b = itemFields.length; a < b; a += 1) { // For each item field
           const currentField = itemFields[a];
           if (this.selectedFields.includes(currentField) === false) { // If selected field includes field, DELETE
@@ -210,37 +206,40 @@ export class Query <ExtendedItem extends Item> {
     this.resultItems = resultItems;
     this.finalized = true;
   }
-  public ids () : string[] {
+
+  ids() {
     if (this.finalized === false) this.finalize();
     return this.resultIds;
   }
-  public items () : Item[] {
+
+  items() {
     if (this.finalized === false) this.finalize();
     return this.resultItems;
   }
-  public firstId () : string|undefined {
+
+  firstId() {
     if (this.finalized === false) this.finalize();
     return this.resultIds[0];
   }
-  public firstItem () : Item|undefined {
+
+  firstItem() {
     if (this.finalized === false) this.finalize();
     return this.resultItems[0];
   }
-  public entries () : [string, Item][] {
+
+  entries() {
     if (this.finalized === false) this.finalize();
-    return this.resultItems.map((item) => [item.id, item] as [string, Item]);
+    return this.resultItems.map(item => [item.id, item]);
   }
-  public results () : [string[], Item[]] {
+
+  results() {
     if (this.finalized === false) this.finalize();
     return [this.resultIds, this.resultItems];
   }
 }
 
-export class Table <ExtendedItem extends Item> {
-  public label: string;
-  private database: Database;
-  public index: Map<string, Item>;
-  public constructor (label: string, database: Database) {
+class Table {
+  constructor(label, database) {
     if (database.initialized === false && database.initializing === false) {
       throw Error('@constructor : Cannot create table on non-initialized database.');
     }
@@ -248,34 +247,37 @@ export class Table <ExtendedItem extends Item> {
     this.database = database;
     this.index = new Map();
     if (database.tables.has(label)) {
-      return database.tables.get(label) as Table<ExtendedItem>;
+      return database.tables.get(label);
     }
     database.tables.set(label, this);
   }
-  public randomItemId () : string {
+
+  randomItemId() {
     let id = uuidv4();
     while (this.index.has(id)) {
       id = uuidv4();
     }
     return id;
   }
-  public async insertItem (id: string, data : ExtendedItem) : Promise<ExtendedItem> {
+
+  async insertItem(id, data) {
     if (isPlainObject(data) === false) {
       throw Error('@insertItem : Invalid "data", "data" must be a plain object');
     }
     if (this.index.has(id)) {
       throw Error('@insertItem : Invalid "id", must not exist in table');
     }
-    const item = {
+    const item = Object.freeze({
       id,
       ...cloneDeep(data),
-    };
+    });
     this.index.set(id, item);
     await this.database.save();
-    const copy = cloneDeep (item);
+    const copy = cloneDeep(item);
     return copy;
   }
-  public async updateItem (modified: ExtendedItem) : Promise<void> {
+
+  async updateItem(modified) {
     if (isPlainObject(modified) === false) {
       throw Error('@updateItem : Invalid "modified", "modified" must be a plain object');
     }
@@ -285,11 +287,12 @@ export class Table <ExtendedItem extends Item> {
     if (this.index.has(modified.id) === false) {
       throw Error('@updateItem : Invalid "item", item "id" must exist in table');
     }
-    const item = cloneDeep  (modified);
-    this.index.set(item.id as string, item);
+    const item = cloneDeep(modified);
+    this.index.set(item.id, item);
     await this.database.save();
   }
-  public async updateItemById (id: string, data: ExtendedItem) : Promise<ExtendedItem> {
+
+  async updateItemById(id, data) {
     if (isPlainObject(data) === false) {
       throw Error('@updateItemById : Invalid "data", "data" must be a plain object');
     }
@@ -305,10 +308,11 @@ export class Table <ExtendedItem extends Item> {
     };
     this.index.set(id, item);
     await this.database.save();
-    const copy = cloneDeep  (item);
+    const copy = cloneDeep(item);
     return copy;
   }
-  public async mergeItemById (id: string, data: ExtendedItem) : Promise<ExtendedItem> {
+
+  async mergeItemById(id, data) {
     if (isPlainObject(data) === false) {
       throw Error('@mergeItemById : Invalid "data", "data" must be a plain object');
     }
@@ -322,14 +326,15 @@ export class Table <ExtendedItem extends Item> {
     const item = {
       id,
       ...existing,
-      ...cloneDeep (data)
+      ...cloneDeep(data),
     };
     this.index.set(id, item);
     await this.database.save();
-    const copy = cloneDeep  (item);
+    const copy = cloneDeep(item);
     return copy;
   }
-  public async removeItem (item: ExtendedItem) : Promise<void> {
+
+  async removeItem(item) {
     if (isPlainObject(item) === false) {
       throw Error('@removeItem : Invalid "item", "item" must be a plain object');
     }
@@ -342,7 +347,8 @@ export class Table <ExtendedItem extends Item> {
     this.index.delete(item.id);
     await this.database.save();
   }
-  public async removeItemById (id: string) : Promise<void> {
+
+  async removeItemById(id) {
     if (id === undefined) {
       throw Error('@removeItemById : Invalid "id", "id" must not be "undefined"');
     }
@@ -352,7 +358,8 @@ export class Table <ExtendedItem extends Item> {
     this.index.delete(id);
     await this.database.save();
   }
-  public fetchItem (id: string) : ExtendedItem {
+
+  fetchItem(id) {
     if (id === undefined) {
       throw Error('@fetchItem : Invalid "id", "id" must not be "undefined"');
     }
@@ -360,28 +367,27 @@ export class Table <ExtendedItem extends Item> {
       throw Error('@fetchItem : Invalid "item", "id" must exist in table');
     }
     const item = this.index.get(id);
-    const copy = cloneDeep  (item as ExtendedItem);
+    const copy = cloneDeep(item);
     return copy;
   }
-  public async clear() : Promise<void> {
+
+  async clear() {
     this.index.clear();
     await this.database.save();
   }
-  public async destroy() : Promise<void> {
+
+  async destroy() {
     this.database.tables.delete(this.label);
     await this.database.save();
   }
-  public query () : Query <ExtendedItem> {
+
+  query() {
     return new Query(this);
   }
 }
 
-
-export class KVTable <Value> {
-  public label: string;
-  private database: Database;
-  public index: Map<string, Value>;
-  public constructor (label: string, database: Database) {
+class KVTable {
+  constructor(label, database) {
     if (database.initialized === false && database.initializing === false) {
       throw Error('@Cannot create KVTable on non-initialized database.');
     }
@@ -389,11 +395,12 @@ export class KVTable <Value> {
     this.database = database;
     this.index = new Map();
     if (database.kvtables.has(label)) {
-      return database.kvtables.get(label) as KVTable<Value>;
+      return database.kvtables.get(label);
     }
     database.kvtables.set(label, this);
   }
-  public async set (key: string, value: Value) : Promise<void> {
+
+  async set(key, value) {
     if (key === undefined) {
       throw Error('@set : Invalid "key", "key" must not be "undefined"');
     }
@@ -403,7 +410,8 @@ export class KVTable <Value> {
     this.index.set(key, value);
     await this.database.save();
   }
-  public has (key: string) : boolean {
+
+  has(key) {
     if (key === undefined) {
       throw Error('@set : Invalid "key", "key" must not be "undefined"');
     }
@@ -412,7 +420,8 @@ export class KVTable <Value> {
     }
     return this.index.has(key);
   }
-  public get (key: string) : Value {
+
+  get(key) {
     if (key === undefined) {
       throw Error('@get : Invalid "key", "key" must not be "undefined"');
     }
@@ -422,9 +431,10 @@ export class KVTable <Value> {
     if (this.index.has(key) === false) {
       throw Error('@get : Invalid "key", "key" must exist in KVTable');
     }
-    return this.index.get(key) as Value;
+    return this.index.get(key);
   }
-  public async delete (key: string) : Promise<void> {
+
+  async delete(key) {
     if (key === undefined) {
       throw Error('@set : Invalid "key", "key" must not be "undefined"');
     }
@@ -434,71 +444,39 @@ export class KVTable <Value> {
     this.index.delete(key);
     await this.database.save();
   }
-  public async clear() : Promise<void> {
+
+  async clear() {
     this.index.clear();
     await this.database.save();
   }
-  public async destroy() : Promise<void> {
+
+  async destroy() {
     this.database.kvtables.delete(this.label);
     await this.database.save();
   }
 }
 
-const rurudbVersion = 7;
-
-// Database
-interface DatabaseOptions {
-  filename: string;
-  directory: string;
-  saveFormat: "json"|"msgpack"|"readable_json";
-  snapshotInterval?: string;
-  msgpackBufferSize?: number;
-  logFunction?: Function;
-}
-
-export class Database {
-  private filename: string;
-  private directory: string;
-  private main: string;
-  private temp: string;
-  private recent: string;
-  private snapshotInterval: number;
-  private lastSnapshotTimestamp: number;
-  public tables: Map<string, Table <Item>>;
-  public kvtables: Map<string, KVTable<unknown>>;
-  private saving: boolean;
-  private queue: [Function, Function][];
-  private mainFd: number;
-  private tempFd: number;
-  private recentFd: number;
-  public initialized: boolean;
-  public initializing: boolean;
-  private saveFormat: "json"|"msgpack"|"readable_json";
-  private msgpackEncode: Function|undefined;
-  private msgpackDecode: Function|undefined;
-  private snapshotExtension: string;
-  private internalInitializePromise: Promise<void>|undefined;
-  private logFunction: Function|undefined;
-  public constructor (options: DatabaseOptions) {
+class Database {
+  constructor(options) {
     this.logFunction = options.logFunction;
     this.filename = options.filename;
     this.directory = options.directory;
     switch (options.saveFormat) {
-      case "json": {
+      case 'json': {
         this.main = options.directory.concat('/', options.filename, '-current.rrdb');
         this.temp = options.directory.concat('/', options.filename, '-temp.rrdb');
         this.recent = options.directory.concat('/', options.filename, '-recent.rrdb');
         this.snapshotExtension = '-snapshot.rrdb';
         break;
       }
-      case "readable_json": {
+      case 'readable_json': {
         this.main = options.directory.concat('/', options.filename, '-current.rrdb');
         this.temp = options.directory.concat('/', options.filename, '-temp.rrdb');
         this.recent = options.directory.concat('/', options.filename, '-recent.rrdb');
         this.snapshotExtension = '-snapshot.rrdb';
         break;
       }
-      case "msgpack": {
+      case 'msgpack': {
         this.main = options.directory.concat('/', options.filename, '-current.prrdb');
         this.temp = options.directory.concat('/', options.filename, '-temp.prrdb');
         this.recent = options.directory.concat('/', options.filename, '-recent.prrdb');
@@ -520,12 +498,12 @@ export class Database {
         }
         break;
       }
-      default:{
+      default: {
         throw Error('@constructor : options.saveFormat must be either "json", "readable_json" or "msgpack"');
       }
     }
     this.saveFormat = options.saveFormat;
-    
+
     this.tables = new Map();
     this.kvtables = new Map();
     this.snapshotInterval = options.snapshotInterval ? ms(options.snapshotInterval) : Infinity;
@@ -541,12 +519,13 @@ export class Database {
     this.initialized = false;
     this.initializing = false;
   }
-  public async initialize () : Promise<void> {
+
+  async initialize() {
     if (this.initialized === false) {
       if (this.internalInitializePromise !== undefined) {
         return this.internalInitializePromise;
       }
-      this.internalInitializePromise = (async () : Promise<void> => {
+      this.internalInitializePromise = (async () => {
         await this.internalLoad();
         this.initialized = true;
       })();
@@ -554,10 +533,11 @@ export class Database {
       this.internalInitializePromise = undefined;
     }
   }
-  private async internalLoad () : Promise<void> {
+
+  async internalLoad() {
     this.tables.clear();
     this.kvtables.clear();
-    let dbDataString: Buffer = Buffer.alloc(0);
+    let dbDataString = Buffer.alloc(0);
     if (await exists(this.directory)) {
       if (await exists(this.main)) {
         this.mainFd = await open(this.main, 'r');
@@ -567,9 +547,7 @@ export class Database {
           const mainContent = Buffer.alloc(mainStat.size);
           await read(this.mainFd, mainContent, 0, mainStat.size, 0);
           dbDataString = mainContent;
-        } else {
-          if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.main} empty, possibly corrupted.`);
-        }
+        } else if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.main} empty, possibly corrupted.`);
         await close(this.mainFd);
       } else {
         if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.main} not found.`);
@@ -581,9 +559,7 @@ export class Database {
             const tempContent = Buffer.alloc(tempStat.size);
             await read(this.tempFd, tempContent, 0, tempStat.size, 0);
             dbDataString = tempContent;
-          } else {
-            if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.temp} empty, possibly corrupted.`);
-          }
+          } else if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.temp} empty, possibly corrupted.`);
           await close(this.tempFd);
         } else {
           if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.temp} not found.`);
@@ -595,25 +571,21 @@ export class Database {
               const tempContent = Buffer.alloc(recentStat.size);
               await read(this.recentFd, tempContent, 0, recentStat.size, 0);
               dbDataString = tempContent;
-            } else {
-              if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.recent} empty, possibly corrupted.`);
-            }
+            } else if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.recent} empty, possibly corrupted.`);
             await close(this.recentFd);
-          } else {
-            if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.recent} not found.`);
-          }
+          } else if (this.logFunction !== undefined) this.logFunction(`@internalLoad : ${this.recent} not found.`);
         }
       }
     }
     if (dbDataString.byteLength !== 0) {
-      let data: any; // eslint-disable-line
-      if (this.saveFormat === "json" || this.saveFormat === "readable_json") {
+      let data; // eslint-disable-line
+      if (this.saveFormat === 'json' || this.saveFormat === 'readable_json') {
         data = JSON.parse(dbDataString.toString());
       } else { // msgpack
-        data = (this.msgpackDecode as Function) (dbDataString);
+        data = (this.msgpackDecode)(dbDataString);
       }
 
-      if (this.logFunction !== undefined) this.logFunction(`@internalLoad : database file loaded, populating tables.`);
+      if (this.logFunction !== undefined) this.logFunction('@internalLoad  file loaded, populating tables.');
       const [dataMeta, dataTables, dataKVTables] = data;
       const [filenameLoaded, rurudbVersionLoaded] = dataMeta;
       if (filenameLoaded !== this.filename) {
@@ -624,27 +596,28 @@ export class Database {
       }
       this.initializing = true;
       for (let i = 0, l = dataTables.length; i < l; i += 1) {
-        const [label, items]: [string, Item[]] = dataTables[i];
-        const table = new Table (label, this);
+        const [label, items] = dataTables[i];
+        const table = new Table(label, this);
         for (let a = 0, b = items.length; a < b; a += 1) {
-          table.index.set(items[a].id as string, items[a]);
+          table.index.set(items[a].id, items[a]);
         }
       }
       for (let i = 0, l = dataKVTables.length; i < l; i += 1) {
-        const [label, keys, values]: [string, unknown[], unknown[]] = dataKVTables[i];
-        const kvtable = new KVTable <unknown> (label, this);
+        const [label, keys, values] = dataKVTables[i];
+        const kvtable = new KVTable(label, this);
         for (let a = 0, b = keys.length; a < b; a += 1) {
-          kvtable.index.set(keys[a] as string, values[a]);
+          kvtable.index.set(keys[a], values[a]);
         }
       }
       this.initializing = false;
-      if (this.logFunction !== undefined) this.logFunction(`@internalLoad : tables populated.`);
+      if (this.logFunction !== undefined) this.logFunction('@internalLoad : tables populated.');
     } else {
-      if (this.logFunction !== undefined) this.logFunction(`@internalLoad : database file not loaded, saving empty db.`);
+      if (this.logFunction !== undefined) this.logFunction('@internalLoad  file not loaded, saving empty db.');
       await this.internalSave();
     }
   }
-  private async internalBeforeSave () : Promise<void> {
+
+  async internalBeforeSave() {
     // Ensure directory existence
     if (await exists(this.directory) === false) {
       await mkdir(this.directory, { recursive: true });
@@ -665,12 +638,12 @@ export class Database {
     this.recentFd = await open(this.recent, 'r+');
     process.nextTick(this.internalSaveLoop);
   }
-  private async internalSaveLoop () : Promise<void> {
+
+  async internalSaveLoop() {
     const fetched = this.queue;
     this.queue = [];
-    let error: Error|undefined = undefined;;
+    let error;
     try {
-
       const tables = Array.from(this.tables.entries());
       const dataTables = new Array(tables.length);
       for (let i = 0, l = tables.length; i < l; i += 1) {
@@ -697,13 +670,13 @@ export class Database {
         dataKVTables,
       ];
 
-      let dataString: Buffer;
-      if (this.saveFormat === "json") {
+      let dataString;
+      if (this.saveFormat === 'json') {
         dataString = Buffer.from(JSON.stringify(data));
-      } else if (this.saveFormat === "readable_json") {
+      } else if (this.saveFormat === 'readable_json') {
         dataString = Buffer.from(JSON.stringify(data, null, 2));
       } else { // msgpack
-        dataString = (this.msgpackEncode as Function) (data) as Buffer;
+        dataString = (this.msgpackEncode)(data);
       }
 
       await ftruncate(this.tempFd);
@@ -713,7 +686,7 @@ export class Database {
       if (mainStat.size > 0) {
         const modified = mainStat.mtime;
         const current = new Date();
-        let mainContent = undefined;;
+        let mainContent;
         if (
           current.valueOf() - modified.valueOf() >= this.snapshotInterval
           && current.valueOf() - this.lastSnapshotTimestamp >= this.snapshotInterval
@@ -724,7 +697,7 @@ export class Database {
             dateToString(current),
             this.snapshotExtension,
           );
-          const snapshotFd = await open(snapshot, 'w');;
+          const snapshotFd = await open(snapshot, 'w');
           mainContent = Buffer.alloc(mainStat.size);
           await read(this.mainFd, mainContent, 0, mainStat.size, 0);
           await write(snapshotFd, mainContent, 0, mainStat.size, 0);
@@ -754,7 +727,7 @@ export class Database {
     }
     if (error !== undefined) {
       for (let i = 0, l = fetched.length; i < l; i += 1) {
-        const [,reject] = fetched[i];
+        const [, reject] = fetched[i];
         reject(error);
       }
     } else {
@@ -764,7 +737,8 @@ export class Database {
       }
     }
   }
-  private async internalSave () : Promise <void> {
+
+  async internalSave() {
     return new Promise((resolve, reject) => {
       this.queue.push([resolve, reject]);
       if (this.saving === false) {
@@ -773,24 +747,34 @@ export class Database {
       }
     });
   }
-  public async save () : Promise<void> {
+
+  async save() {
     if (this.initialized === false) {
       throw Error('@save : Cannot call "save" on non-initialized database.');
     }
     await this.internalSave();
   }
-  public table (label: string) : Table <Item> {
+
+  table(label) {
     if (this.tables.has(label) === false) {
       throw Error(`@table : "${label}" Table not found, must exist`);
     }
-    const table = new Table (label, this);
+    const table = new Table(label, this);
     return table;
   }
-  public kvtable (label: string) : KVTable <unknown> {
+
+  kvtable(label) {
     if (this.kvtables.has(label) === false) {
       throw Error(`@kvtable : "${label}" KVTable not found, must exist`);
     }
-    const kvtable = new KVTable (label, this);
+    const kvtable = new KVTable(label, this);
     return kvtable;
   }
 }
+
+module.exports = {
+  Database,
+  Query,
+  Table,
+  KVTable,
+};
