+++ jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
-  preset: 'ts-jest',
   testEnvironment: 'node',
   collectCoverage: true,
-  collectCoverageFrom : ['src/*.ts']
+  collectCoverageFrom : ['./index.js']
 };
