--- src/@types/tinydate/index.d.ts
@@ -1,4 +1,0 @@
-declare module 'tinydate' {
-  function tinydate (pattern:string) : ((date: Date) => string);
-  export = tinydate;
-}
