--- src/@types/what-the-pack/index.d.ts
@@ -1,8 +1,0 @@
-declare module 'what-the-pack' {
-  interface MessagePackFunctions {
-    encode: (value: any) => Buffer; // eslint-disable-line
-    decode: (buffer: Buffer) => any; // eslint-disable-line
-    register: (...strings: string[]) => void;
-  }
-  export const initialize: (tempBufferLength: number, logFunction ?: Function) => MessagePackFunctions;
-}
